Validate field names match DCM
