+++ azext_iot/central/providers/device_provider.py
@@ -4,7 +4,7 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-from azext_iot.services import central
+from azext_iot.central import services as central_services
 
 
 class CentralDeviceProvider:
@@ -36,7 +36,7 @@ class CentralDeviceProvider:
         ):
             self._device_templates[
                 device_template_urn
-            ] = central.device_template.get_device_template(
+            ] = central_services.device_template.get_device_template(
                 cmd, device_template_urn, app_id, token, central_dns_suffix
             )
 
@@ -60,7 +60,7 @@ class CentralDeviceProvider:
             raise ValueError("Device id must be specified.")
 
         if device_id not in self._devices or not self._devices.get(device_id):
-            self._devices[device_id] = central.device.get_device(
+            self._devices[device_id] = central_services.device.get_device(
                 cmd, device_id, app_id, token, central_dns_suffix
             )
