+++ azext_iot/central/services/_utility.py
@@ -3,7 +3,14 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
+# Nothing in this file should be used outside of service/central
 
-from . import central
+from azext_iot.common import auth
 
-__all__ = ["central"]
+
+def get_token(token, cmd):
+    if not token:
+        token = auth.get_aad_token(cmd, resource="https://apps.azureiotcentral.com")
+        return "Bearer {}".format(token["accessToken"])
+
+    return token
