+++ azext_iot/central/services/device_template.py
