+++ azext_iot/common/auth.py
