+++ azext_iot/operations/central.py
@@ -13,7 +13,7 @@ from azext_iot.common._azure import get_iot_hub_token_from_central_app_id
 from azext_iot.common.shared import SdkType
 from azext_iot.common.utility import unpack_msrest_error, init_monitoring
 from azext_iot.common.sas_token_auth import BasicSasTokenAuthentication
-from azext_iot.providers.central import CentralDeviceProvider
+from azext_iot.central.providers import CentralDeviceProvider
 
 
 def find_between(s, start, end):
